// src/app/thirparty/model/apimodel.ts
import { animate, style, transition, trigger } from '@angular/animations';

export class usermodel {
  source: any;
  requestip: any;
  Browser: any;
  hskvalue: any;
  browsername: any;
  longitude: any;
  latitude: any;
  gioaddress: any;
}
export interface Person {
  param1: any;
  param2: any;
  param3: any;
}
export class play_field_data {
  COURT_NAME: any;
  TEXT_AREA: any;
  PLAY_FIELD: any;
  REQUIRED_DETAILS: any;
  PLAY_FILED_CODE: any;
  CENTS: any;
  TEXT_AREA_NAME: any;
  CENTS_NAME: any;
}

export class fund_brekdown {
  description: any;
  amount: any;
}

export class basemodel {
  type: any;
  param1: any;
  param2: any;
  param3: any;
  param4: any;
  param5: any;
  param6: any;
  param7: any;
  param8: any;
  param9: any;
  param10: any;
  param11: any;
  param12: any;
  param13: any;
  param14: any;
  param15: any;
  param16: any;
  param17: any;
  param18: any;
  param19: any;
  param20: any;
  param21: any;
  param22: any;
  param23: any;
  param24: any;
  param25: any;
  param26: any;
  param27: any;
  param28: any;
  param29: any;
  param30: any;
  param31: any;
  param32: any;
  param33: any;
  param34: any;
  param35: any;
  param36: any;
  param37: any;
  param38: any;
  param39: any;
  param40: any;
  json1: any;
  json2: any;
  json3: any;
  json4: any;
  remarks: any;
  rucode: any;
  usersno: any;
  roleid: any;
  source: any;
  requestip: any;
  Browser: any;
  hskvalue: any;
  browsername: any;
  longitude: any;
  latitude: any;
  gioaddress: any;
  refno: any;
  modulename: any;
  islogstore: any;
  logfoldername: any;
  signuser: any;
}